osa 4 exercises: blogilista

+++ osa4/blogilista/controllers/blogs.js
@@ -1,10 +1,13 @@
 const blogsRouter = require('express').Router()
 const { request } = require('../app')
-const blog = require('../models/blog')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs)
 })
 
@@ -19,22 +22,38 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
 
-  /*if (!body.title || !body.url) {
-    return response.status(400).json({error: 'title or url missing'})
-  }*/
+  if (!body.title) {
+    return response.status(400).json({ error: 'title is missing' })
+  }
+
+  if (!body.url) {
+    return response.status(400).json({ error: 'url is missing' })
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
@@ -42,10 +61,9 @@ blogsRouter.delete('/:id', async (request, response) => {
 blogsRouter.put('/:id', async (request, response) => {
   const blog = await Blog.findById(request.params.id)
   if (blog) {
-    blog.likes = likes
+    blog.likes = request.body.likes
     const updatedBlog = await blog.save()
     response.status(200).json(updatedBlog)
-    response.json(blog)
   } else {
     response.status(404).end()
   }
